+++ cfgov/v1/atomic_elements/molecules.py
@@ -387,38 +387,3 @@ class ContentWithAnchor(blocks.StructBlock):
     class Meta:
         icon = 'edit'
         template = '_includes/molecules/full-width-text-anchor.html'
-
-
-class SnippetList(blocks.StructBlock):
-    heading = blocks.CharBlock(required=False)
-    body = blocks.RichTextBlock(required=False)
-    image = atoms.ImageBasic(required=False)
-
-    snippet_type = blocks.ChoiceBlock(
-        choices=ref.snippet_types,
-        required=True
-    )
-    actions = blocks.ListBlock(blocks.StructBlock([
-        ('link_label', blocks.CharBlock(
-            help_text='E.g., "Download" or "Order free prints"'
-        )),
-        ('snippet_field', blocks.ChoiceBlock(
-            choices=ref.snippet_fields,
-            help_text='Corresponds to the available fields for the selected'
-                      'snippet type.'
-        )),
-    ]))
-
-    # Needs a tag chooser
-
-    tags = blocks.ListBlock(
-        blocks.CharBlock(label='Tag'),
-        help_text='Enter tag names to filter the snippets. For a snippet to '
-                  'match and be output in the list, it must have been tagged '
-                  'with all of the tag names listed here. The tag names '
-                  'are case-insensitive.'
-    )
-
-    class Meta:
-        icon = 'table'
-        template = '_includes/molecules/snippet-list.html'